+++ source/calc/dataseries/dailydataseriecalc.cpp
@@ -164,7 +164,7 @@ void DailyDataSerieCalc::serieToStream()
     out << m_assetId.name;
     out.setVersion(QDataStream::Qt_6_0);
     DataSerieValue dt;
-    for(size_t i{0}; i < ar_values.count(); i++)
+    for(long long i{0}; i < ar_values.count(); i++)
     {
         dt = *ar_values[i];
         out << dt.dOpen;
