+++ source/dataseries/customdataserie.cpp
@@ -19,7 +19,7 @@ void CustomDataSerie::serieToStream()
     out << m_assetId.name;
     out.setVersion(QDataStream::Qt_6_0);
     DataSerieValue dt;
-    for(size_t i{0}; i < ar_values.count(); i++)
+    for(long long i{0}; i < ar_values.count(); i++)
     {
         dt = *ar_values[i];
         out << dt.dOpen;
@@ -61,7 +61,7 @@ void CustomDataSerie::loadSerieFromJsonAV(QString json)
 
 void CustomDataSerie::ClearDataSerie()
 {
-    for(size_t i{0}; i < ar_values.count(); i++)
+    for(long long i{0}; i < ar_values.count(); i++)
     {
         delete ar_values[i];
     }
