+++ source/indicatormanager.cpp
@@ -34,7 +34,7 @@ void IndicatorManager::addNewIndicatorData(IndicatorIdentifier id, size_t start,
         series->arr = new DoublyArray();
         series->arr->resize(array.count());
         for(size_t i{start}; i < size; i++){
-            for(size_t j{0}; j < series->arr->count(); ++j){
+            for(long long j{0}; j < series->arr->count(); ++j){
                 (*series->arr)[j].append(array[j][i]);
             }
         }
