+++ source/indicators/movingaverage.cpp
@@ -12,7 +12,7 @@ void MovingAverage::addNewValue(size_t start, size_t count, DoublyArray* values)
     else if (count > Size()) grow(Size() - count);
 
     for(size_t i{start}; i < count; ++i){
-        for(size_t j{0}; j < m_arData.count(); ++j){
+        for(long long j{0}; j < m_arData.count(); ++j){
             m_arData[j][i] = (*values)[j][i - start];
         }
     }
